APIs created with placeholder comments instead of firebase call
